Update firmware to 32bit encoder counter and add stop and reset to binary commands

+++ epuck_firmware/library/motor_led/advance_one_timer/e_motors.h
@@ -58,10 +58,10 @@ void e_set_speed_left(int motor_speed);  // motor speed: from -1000 to 1000
 void e_set_speed_right(int motor_speed); // motor speed: from -1000 to 1000
 void e_set_speed(int linear_speed, int angular_speed);
 
-void e_set_steps_left(int steps_left);
-void e_set_steps_right(int steps_right);
+void e_set_steps_left(long steps_left);
+void e_set_steps_right(long steps_right);
 
-int e_get_steps_left();
-int e_get_steps_right();
+long e_get_steps_left();
+long e_get_steps_right();
 
 #endif
